test: Prepare to split post tests.
